Add huahaha command
Adding playback volume configuration on the config file to ease the
volume change in the future.

It also change the way the audio file played. Now it uses the opus audio
stream directly instead of using ffmpeg library to encode it. This
change is made to make the maximize the bot performance.